+++ resources/views/admin/blocks/types/content_header.blade.php
@@ -6,7 +6,7 @@
     $isNonDefaultLocale = isset($activeLocale) && ! $isDefaultLocale;
 @endphp
 
-@once
+@once('webblocks-builder-items-script')
     @push('admin-scripts')
         @include('webblocks-cms::admin.partials.admin-script', ['path' => 'cms/js/admin/builder-items.js'])
     @endpush