+++ resources/views/admin/pages/partials/slot-block-modal-fragment.blade.php
@@ -1,0 +1,5 @@
+@include('webblocks-cms::admin.pages.partials.slot-block-modal')
+
+<template data-wb-slot-block-fragment-overlays>
+    @stack('overlays')
+</template>